Update SHARED_GEOMETRY_GROUPS test.
