+++ shared-data/js/__tests__/labwareDefSchemaV2.test.ts
@@ -4,7 +4,12 @@ import Ajv from 'ajv'
 import { describe, expect, it, beforeAll, test } from 'vitest'
 
 import schema from '../../labware/schemas/2.json'
-import type { LabwareDefinition2, LabwareWell } from '../types'
+import type {
+  InnerWellGeometry,
+  LabwareDefinition2,
+  LabwareWell,
+} from '../types'
+import { SHARED_GEOMETRY_GROUPS } from './sharedGeometryGroups'
 
 const definitionsDir = path.join(__dirname, '../../labware/definitions/2')
 const fixturesDir = path.join(__dirname, '../../labware/fixtures/2')
@@ -319,3 +324,83 @@ describe('test schemas of all v2 labware fixtures', () => {
     expectGroupsFollowConvention(labwareDef, filename)
   })
 })
+
+describe('check groups of labware that should have the same geometry', () => {
+  describe.each(
+    Object.entries(SHARED_GEOMETRY_GROUPS).map(([groupName, groupEntries]) => ({
+      groupName,
+      groupEntries,
+    }))
+  )('$groupName', ({ groupEntries }) => {
+    const normalizedGroupEntries = groupEntries.map(entry => ({
+      loadName: typeof entry === 'string' ? entry : entry.loadName,
+      geometryKey: typeof entry === 'string' ? undefined : entry.geometryKey,
+    }))
+    test.each(normalizedGroupEntries)(
+      '$loadName',
+      ({ loadName, geometryKey }) => {
+        // We arbitrarily pick the first labware in the group to compare the rest against.
+        const otherLabwareGeometry = getGeometry(
+          normalizedGroupEntries[0].loadName,
+          normalizedGroupEntries[0].geometryKey
+        )
+        const thisLabwareGeometry = getGeometry(loadName, geometryKey)
+        expect(thisLabwareGeometry).toEqual(otherLabwareGeometry)
+      }
+    )
+  })
+})
+
+/**
+ * Return the latest version of the given labware that's defined in schema 2.
+ *
+ * todo(mm, 2025-02-27): We already have a "production" getLatestLabwareDef() function
+ * elsewhere, and it would be nice to reuse that, but it looks like that one currently
+ * relies on a hard-coded list of labware.
+ */
+function findLatestDefinition(loadName: string): LabwareDefinition2 {
+  const candidates: LabwareDefinition2[] = glob
+    .sync('*.json', {
+      cwd: path.join(definitionsDir, loadName),
+      absolute: true,
+    })
+    .map(require)
+  if (candidates.length === 0) {
+    throw new Error(`No definitions found for ${loadName}.`)
+  }
+  candidates.sort((a, b) => a.version - b.version)
+  const latest = candidates[candidates.length - 1]
+  return latest
+}
+
+/**
+ * Extract the given geometry from the given definition.
+ *
+ * If geometryKey is unspecified, the definition is expected to have exactly one
+ * geometry key, and that one is extracted and returned.
+ */
+function getGeometry(
+  loadName: string,
+  geometryKey: string | undefined
+): InnerWellGeometry {
+  const definition = findLatestDefinition(loadName)
+  const availableGeometries = definition.innerLabwareGeometry ?? {}
+
+  if (geometryKey === undefined) {
+    const availableGeometryEntries = Object.entries(availableGeometries)
+    if (availableGeometryEntries.length !== 1) {
+      throw new Error(
+        `Expected exactly 1 geometry in ${definition.parameters.loadName} but found ${availableGeometryEntries.length}.`
+      )
+    }
+    return availableGeometryEntries[0][1]
+  } else {
+    const result = availableGeometries[geometryKey]
+    if (result === undefined) {
+      throw new Error(
+        `No geometry found in ${definition.parameters.loadName} with key ${geometryKey}.`
+      )
+    }
+    return result
+  }
+}
